+++ clap_builder/src/util/id.rs
@@ -172,6 +172,21 @@ impl PartialEq<Id> for String {
     }
 }
 
+#[cfg(feature = "serde")]
+impl serde::Serialize for Id {
+    fn serialize<S: serde::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
+        self.as_str().serialize(serializer)
+    }
+}
+
+#[cfg(feature = "serde")]
+impl<'de> serde::Deserialize<'de> for Id {
+    fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
+        let s = String::deserialize(deserializer)?;
+        Ok(Id::from(s))
+    }
+}
+
 #[cfg(test)]
 #[cfg(feature = "string")]
 mod tests {